fix: explanation in operator-overloading.md (#1881)
This matches what's actually generated by the compiler, at least for the JVM backend (didn't verify on other backends). Using `a.inc()` would be incorrect because `a` could be an expression with side-effects and hence would be evaluated twice: Once during initialization of `a0`, and then again in that step.